update response CRUD + handle exception

+++ src/main/java/vn/taidung/springsocial/service/PostService.java
@@ -10,6 +10,7 @@ import vn.taidung.springsocial.model.Post;
 import vn.taidung.springsocial.model.request.CreatePostRequest;
 import vn.taidung.springsocial.model.request.UpdatePostRequest;
 import vn.taidung.springsocial.repository.PostRepository;
+import vn.taidung.springsocial.util.errors.PostNotFoundException;
 
 @Service
 public class PostService {
@@ -27,15 +28,23 @@ public class PostService {
         return this.postRepository.save(post);
     }
 
-    public Optional<Post> getPostHandler(Long id) {
-        return this.postRepository.findById(id);
+    public Post getPostHandler(Long id) {
+        Optional<Post> optionalPost = this.postRepository.findById(id);
+        if (optionalPost.isPresent()) {
+            return optionalPost.get();
+        }
+        return null;
     }
 
-    public void deletePostHandler(Long id) {
+    public void deletePostHandler(Long id) throws PostNotFoundException {
+        if (!this.postRepository.existsById(id)) {
+            throw new PostNotFoundException("post not found");
+        }
+
         this.postRepository.deleteById(id);
     }
 
-    public Post updatePostHandler(Long id, UpdatePostRequest updatePostRequest) {
+    public Post updatePostHandler(Long id, UpdatePostRequest updatePostRequest) throws PostNotFoundException {
         Optional<Post> optionalPost = this.postRepository.findById(id);
         if (optionalPost.isPresent()) {
             Post post = this.modelMapper.map(optionalPost.get(), Post.class);
@@ -48,8 +57,9 @@ public class PostService {
                 post.setContent(updatePostRequest.getContent());
             }
             return this.postRepository.save(post);
+        } else {
+            throw new PostNotFoundException("Post not found");
         }
-        return null;
     }
 
 }
